force int64

+++ ffsi/optimize_galahad.py
@@ -190,8 +190,11 @@ def optimize(G, I_data, I_data_std, sigma=None):
     m_c = len(p_dims)
 
     # set GALAHAD SNLS cohorts
+    # GALAHAD's Python interface is built with 64-bit integers and rejects any
+    # other int width; numpy's default int is 32-bit on Windows (LLP64), so
+    # cast the index/cohort arrays to int64 explicitly for cross-platform safety.
     ch_list = [i * np.ones(n, dtype=int) for i,n in enumerate(p_dims)]
-    cohort = np.concat(( np.array([-1,-1]), *ch_list))
+    cohort = np.concat((np.array([-1, -1]), *ch_list)).astype(np.int64)
 
     # set GALAHAD SNLS Jacobian info
     if sigma is None: # no regularization
@@ -216,9 +219,9 @@ def optimize(G, I_data, I_data_std, sigma=None):
         for st, dim in zip(starts, p_dims):
             Jr_reg1_col.append(np.arange(st+1, st+dim))
             Jr_reg2_col.append(np.arange(st, st+dim-1))
-        # combined derivative
-        Jr_row = np.concat((Jr_eps_row,*Jr_reg1_row,*Jr_reg2_row))
-        Jr_col = np.concat((Jr_eps_col,*Jr_reg1_col,*Jr_reg2_col))
+        # combined derivative (int64 for GALAHAD; see cohort note above)
+        Jr_row = np.concat((Jr_eps_row, *Jr_reg1_row, *Jr_reg2_row)).astype(np.int64)
+        Jr_col = np.concat((Jr_eps_col, *Jr_reg1_col, *Jr_reg2_col)).astype(np.int64)
     Jr_ptr_ne = 0
     Jr_ptr = None
